Merging fix for IDENTITY-2888 and IDENTITY-2869

+++ core/org.wso2.carbon.user.core/src/main/java/org/wso2/carbon/user/core/common/AbstractUserStoreManager.java
@@ -806,7 +806,16 @@ public abstract class AbstractUserStoreManager implements UserStoreManager {
 		if (attributeName == null || attributeName.isEmpty()) {
 			attributeName = claimManager.getAttributeName(claimURI);
 		}
-		return attributeName != null ? attributeName : claimURI;
+
+		if (attributeName == null) {
+			if (UserCoreConstants.PROFILE_CONFIGURATION.equals(claimURI)) {
+				attributeName = claimURI;
+			} else {
+				throw new UserStoreException("Invalid claim URI value");
+			}
+		}
+
+		return attributeName;
 	}
 
 	/**
@@ -919,6 +928,10 @@ public abstract class AbstractUserStoreManager implements UserStoreManager {
 		}
 		// #################### </Listeners> #####################################################
 
+		if (!doCheckExistingUser(userName)) {
+			throw new UserStoreException("User does not exist. Username : " + userName);
+		}
+
 		doSetUserClaimValue(userName, claimURI, claimValue, profileName);
 
 		// #################### <Listeners> #####################################################
@@ -957,6 +970,10 @@ public abstract class AbstractUserStoreManager implements UserStoreManager {
 			}
 		}
 		// #################### </Listeners> #####################################################
+
+		if (!doCheckExistingUser(userName)) {
+			throw new UserStoreException("User does not exist. Username : " + userName);
+		}
 
 		doSetUserClaimValues(userName, claims, profileName);
 
@@ -997,6 +1014,10 @@ public abstract class AbstractUserStoreManager implements UserStoreManager {
 		}
 		// #################### </Listeners> #####################################################
 
+		if (!doCheckExistingUser(userName)) {
+			throw new UserStoreException("User does not exist. Username : " + userName);
+		}
+
 		doDeleteUserClaimValue(userName, claimURI, profileName);
 
 		// #################### <Listeners> #####################################################
@@ -1034,6 +1055,10 @@ public abstract class AbstractUserStoreManager implements UserStoreManager {
 			}
 		}
 		// #################### </Listeners> #####################################################
+
+		if (!doCheckExistingUser(userName)) {
+			throw new UserStoreException("User does not exist. Username : " + userName);
+		}
 
 		doDeleteUserClaimValues(userName, claims, profileName);
